extract RestApiRequest interface
