+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v1/beacon/GetStateForkTest.java
@@ -28,6 +28,7 @@ import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.beaconrestapi.AbstractMigratedBeaconHandlerTest;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequest;
+import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequestImpl;
 import tech.pegasys.teku.spec.datastructures.metadata.StateAndMetaData;
 
 public class GetStateForkTest extends AbstractMigratedBeaconHandlerTest {
@@ -50,7 +51,7 @@ public class GetStateForkTest extends AbstractMigratedBeaconHandlerTest {
   public void shouldReturnForkInfo() throws Exception {
     when(chainDataProvider.getBeaconStateAndMetadata(eq("head")))
         .thenReturn(SafeFuture.completedFuture(Optional.of(stateAndMetaData)));
-    request = new RestApiRequest(context, handler.getMetadata());
+    request = new RestApiRequestImpl(context, handler.getMetadata());
 
     handler.handleRequest(request);
 
@@ -64,7 +65,7 @@ public class GetStateForkTest extends AbstractMigratedBeaconHandlerTest {
   public void shouldReturnNotFound() throws Exception {
     when(chainDataProvider.getBeaconStateAndMetadata(eq("head")))
         .thenReturn(SafeFuture.completedFuture(Optional.empty()));
-    request = new RestApiRequest(context, handler.getMetadata());
+    request = new RestApiRequestImpl(context, handler.getMetadata());
 
     handler.handleRequest(request);
