+++ infrastructure/restapi/src/main/java/tech/pegasys/teku/infrastructure/restapi/endpoints/JavalinEndpointAdapter.java
@@ -32,7 +32,7 @@ public class JavalinEndpointAdapter implements Handler {
 
   @Override
   public void handle(final Context ctx) throws Exception {
-    final RestApiRequest request = new RestApiRequest(ctx, endpoint.getMetadata());
+    final RestApiRequest request = new RestApiRequestImpl(ctx, endpoint.getMetadata());
     endpoint.handleRequest(request);
   }
 }
