+++ infrastructure/restapi/src/main/java/tech/pegasys/teku/infrastructure/restapi/endpoints/RestApiRequest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 ConsenSys AG.
+ * Copyright 2022 ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -13,150 +13,32 @@
 
 package tech.pegasys.teku.infrastructure.restapi.endpoints;
 
-import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_INTERNAL_SERVER_ERROR;
-import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
-import static tech.pegasys.teku.infrastructure.http.RestApiConstants.HEADER_ACCEPT;
-import static tech.pegasys.teku.infrastructure.restapi.endpoints.BadRequest.BAD_REQUEST_TYPE;
-
 import com.fasterxml.jackson.core.JsonProcessingException;
-import io.javalin.core.util.Header;
-import io.javalin.http.Context;
-import java.io.ByteArrayInputStream;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
-import tech.pegasys.teku.infrastructure.http.HttpErrorResponse;
-import tech.pegasys.teku.infrastructure.json.JsonUtil;
 
-public class RestApiRequest {
-  private static final Logger LOG = LogManager.getLogger();
-  private final Context context;
-  private final EndpointMetadata metadata;
-  private final Map<String, String> pathParamMap;
-  private final Map<String, List<String>> queryParamMap;
-
+public interface RestApiRequest {
   @SuppressWarnings({"TypeParameterUnusedInFormals"})
-  public <T> T getRequestBody() throws JsonProcessingException {
-    return metadata.getRequestBody(
-        context.bodyAsInputStream(), Optional.ofNullable(context.header("Content-Type")));
-  }
+  <T> T getRequestBody() throws JsonProcessingException;
 
-  public RestApiRequest(final Context context, final EndpointMetadata metadata) {
-    this.context = context;
-    this.metadata = metadata;
-    this.pathParamMap = context.pathParamMap();
-    this.queryParamMap = context.queryParamMap();
-  }
+  void respondOk(Object response) throws JsonProcessingException;
 
-  public void respondOk(final Object response) throws JsonProcessingException {
-    respond(SC_OK, response);
-  }
+  void respondAsync(SafeFuture<AsyncApiResponse> futureResponse);
 
-  public void respondAsync(final SafeFuture<AsyncApiResponse> futureResponse) {
-    context.future(
-        futureResponse
-            .thenApply(
-                result -> {
-                  try {
-                    return respond(result.getResponseCode(), result.getResponseBody());
-                  } catch (JsonProcessingException e) {
-                    LOG.trace("Failed to generate API response", e);
-                    context.status(SC_INTERNAL_SERVER_ERROR);
-                    return Bytes.EMPTY.toArrayUnsafe();
-                  }
-                })
-            .thenApply(ByteArrayInputStream::new));
-  }
+  void respondOk(Object response, CacheLength cacheLength) throws JsonProcessingException;
 
-  public void respondOk(final Object response, final CacheLength cacheLength)
-      throws JsonProcessingException {
-    context.header(Header.CACHE_CONTROL, cacheLength.getHttpHeaderValue());
-    respond(SC_OK, response);
-  }
+  void respondError(int statusCode, String message) throws JsonProcessingException;
 
-  public void respondError(final int statusCode, final String message)
-      throws JsonProcessingException {
-    respond(statusCode, new HttpErrorResponse(statusCode, message));
-  }
+  void respondWithCode(int statusCode);
 
-  private byte[] respond(final int statusCode, final Optional<Object> response)
-      throws JsonProcessingException {
-    context.status(statusCode);
-    if (response.isPresent()) {
-      return respondImpl(statusCode, response.get());
-    }
-    return Bytes.EMPTY.toArrayUnsafe();
-  }
+  <T> T getPathParameter(ParameterMetadata<T> parameterMetadata);
 
-  private void respond(final int statusCode, final Object response) throws JsonProcessingException {
-    context.status(statusCode);
-    final byte[] responseData = respondImpl(statusCode, response);
-    context.result(responseData);
-  }
+  <T> Optional<T> getOptionalQueryParameter(ParameterMetadata<T> parameterMetadata);
 
-  private byte[] respondImpl(final int statusCode, final Object response)
-      throws JsonProcessingException {
-    final ResponseMetadata responseMetadata =
-        metadata.createResponseMetadata(
-            statusCode, Optional.ofNullable(context.header(HEADER_ACCEPT)), response);
-    context.contentType(responseMetadata.getContentType());
-    responseMetadata.getAdditionalHeaders().forEach(context::header);
-    return metadata.serialize(statusCode, responseMetadata.getContentType(), response);
-  }
+  <T> T getQueryParameter(ParameterMetadata<T> parameterMetadata);
 
-  /** This is only used when intending to return status code without a response body */
-  public void respondWithCode(final int statusCode) {
-    context.status(statusCode);
-  }
-
-  public <T> T getPathParameter(final ParameterMetadata<T> parameterMetadata) {
-    return parameterMetadata
-        .getType()
-        .deserializeFromString(pathParamMap.get(parameterMetadata.getName()));
-  }
-
-  public <T> Optional<T> getOptionalQueryParameter(final ParameterMetadata<T> parameterMetadata) {
-    if (!queryParamMap.containsKey(parameterMetadata.getName())) {
-      return Optional.empty();
-    }
-    return Optional.of(
-        parameterMetadata
-            .getType()
-            .deserializeFromString(
-                SingleQueryParameterUtils.validateQueryParameter(
-                    queryParamMap, parameterMetadata.getName())));
-  }
-
-  public <T> T getQueryParameter(final ParameterMetadata<T> parameterMetadata) {
-    return parameterMetadata
-        .getType()
-        .deserializeFromString(
-            SingleQueryParameterUtils.validateQueryParameter(
-                queryParamMap, parameterMetadata.getName()));
-  }
-
-  public <T> void handleOptionalResult(
-      SafeFuture<Optional<T>> future, ResultProcessor<T> resultProcessor, final int missingStatus) {
-    context.future(
-        future.thenApplyChecked(
-            result -> {
-              if (result.isPresent()) {
-                return resultProcessor.process(context, result.get()).orElse(null);
-              } else {
-                context.status(missingStatus);
-                return JsonUtil.serialize(
-                    new BadRequest(missingStatus, "Not found"), BAD_REQUEST_TYPE);
-              }
-            }));
-  }
-
-  @FunctionalInterface
-  public interface ResultProcessor<T> {
-    // Process result, returning an optional serialized response
-    Optional<String> process(Context context, T result) throws Exception;
-  }
+  <T> void handleOptionalResult(
+      SafeFuture<Optional<T>> future,
+      RestApiRequestImpl.ResultProcessor<T> resultProcessor,
+      int missingStatus);
 }
